refactor database mock
